MelodyGenerator refactor - start - separated class
